requirements.txt
psycopg2-binary==2.8.6
aiogram==2.12